requirements.txt
# Core dependencies
yt-dlp>=2024.1.0
gallery-dl>=1.31.0
faster-whisper>=1.0.0
ollama>=0.1.0
rich>=13.7.0
playwright>=1.40.0

# Utilities
python-dateutil>=2.8.2
youtube-comment-downloader
psutil>=5.9.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# RAG / Vector search
chromadb>=0.3.27
sentence-transformers>=2.2.2
langchain-text-splitters>=0.3.0

# MCP / Server
fastapi>=0.95.0
uvicorn>=0.22.0
mcp[sse]>=0.1.0
python-multipart>=0.0.6


# JWT support for MCP authentication
PyJWT>=2.8.0

